move vault get method to entity
