+++ src/mappings/spacewalk.ts
@@ -1,13 +1,8 @@
 import { EventHandlerContext } from '../processor'
-import { Vault, RedeemRequest, IssueRequest } from '../model'
+import { RedeemRequest, IssueRequest } from '../model'
 import { beautifyCurrencyIdString } from '../mappings/token'
 import { hexToSs58 } from './nabla/addresses'
-
-interface VaultIdFlat {
-    accountId: string
-    collateral: string
-    wrapped: string
-}
+import { VaultIdFlat, getOrCreateVault } from '../entities/vault'
 
 export async function handleIssueRequest(ctx: EventHandlerContext) {
     const { id: eventId, args } = ctx.event
@@ -51,28 +46,6 @@ export async function handleRedeemRequest(ctx: EventHandlerContext) {
     })
 
     await ctx.store.save(redeemRequest)
-}
-
-export async function getOrCreateVault(
-    ctx: EventHandlerContext,
-    vaultId: VaultIdFlat,
-    vaultStellarPublicKey: string
-): Promise<Vault | undefined> {
-    const vaultIdString = `${vaultId.accountId}-${vaultId.collateral}-${vaultId.wrapped}`
-    let vault = await ctx.store.get(Vault, vaultIdString)
-
-    if (!vault) {
-        vault = new Vault({
-            id: vaultIdString,
-            accountId: hexToSs58(vaultId.accountId),
-            collateral: vaultId.collateral,
-            wrapped: vaultId.wrapped,
-            vaultStellarPublicKey: vaultStellarPublicKey,
-        })
-        await ctx.store.save(vault)
-    }
-
-    return vault
 }
 
 function parseVaultId(vaultId: any): VaultIdFlat {
